Created base class for exaalt, moved BUILDLammps to lammps_base.py
